refactor: refactor types
