+++ src/components/ListMovies/ListMovies.tsx
@@ -2,14 +2,7 @@ import React from 'react';
 import {FlatList, StyleSheet, View} from 'react-native';
 import {MovieCard} from '../MovieCard/MovieCard';
 import {useSafeAreaInsets} from 'react-native-safe-area-context';
-
-export type dataType = {
-  Poster: string;
-  Title: string;
-  imdbID: string;
-};
-
-export type ListMoviesType = Array<dataType>;
+import {ListMoviesType} from '../types/moviesTypes';
 
 type ListProps = {
   data: ListMoviesType;
