+++ src/components/Navigation/SearchStack.tsx
@@ -3,8 +3,9 @@ import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import DetailsScreen from '../../page/DetailsScreen/DetailsScreen';
 import SearchScreen from '../../page/SearchScreen/SearchScreen';
 import constants from '../../styles/constants';
+import {SearchStackParamList} from '../types/navigationTypes';
 
-const Stack = createNativeStackNavigator();
+const Stack = createNativeStackNavigator<SearchStackParamList>();
 
 function SearchStack() {
   return (
