+++ src/components/types/moviesTypes.ts
@@ -1,0 +1,7 @@
+export type MoviesType = {
+  Poster: string;
+  Title: string;
+  imdbID: string;
+};
+
+export type ListMoviesType = Array<MoviesType>;
